+++ deepxube/pathfinding/greedy_policy_q.py
@@ -1,7 +1,7 @@
 from typing import List, Any, Type, Optional, TypeVar, Dict
 from abc import ABC
 from deepxube.base.factory import Parser
-from deepxube.base.domain import Domain, ActsEnum, State, Goal
+from deepxube.base.domain import Domain, ActsEnum, State, Action, Goal
 from deepxube.base.pathfinding import InstanceQ, Node, PathFindQHeur, EdgeQ, PathFindQExpandEnum
 from deepxube.pathfinding.utils.search import greedy_next_idx
 from deepxube.factories.pathfinding_factory import pathfinding_factory
@@ -11,7 +11,7 @@ import time
 class InstanceGrPolQ(InstanceQ):
     def __init__(self, root_node: Node, temp: float, eps: float, inst_info: Any):
         super().__init__(root_node, inst_info)
-        self.curr_edge: EdgeQ = EdgeQ(self.root_node, None, self.root_node.heuristic)
+        self.curr_node: Node = self.root_node
         self.temp: float = temp
         assert (eps >= 0.0) and (eps <= 1.0)
         self.eps: float = eps
@@ -37,7 +37,10 @@ class GreedyPolicyQActsAny(PathFindQHeur[D, InstanceGrPolQ], ABC):
 
     def make_instances(self, states: List[State], goals: List[Goal], inst_infos: Optional[List[Any]] = None, compute_root_heur: bool = True,
                        temp: Optional[float] = None, eps: Optional[float] = None) -> List[InstanceGrPolQ]:
-        nodes_root: List[Node] = self._create_root_nodes(states, goals, compute_root_heur=compute_root_heur)
+        """ Always computes the root heurisitc always (regardless of argument)
+
+        """
+        nodes_root: List[Node] = self._create_root_nodes(states, goals, compute_root_heur=True)
         temp_inst: float = temp if temp is not None else self.temp_default
         eps_inst: float = eps if eps is not None else self.eps_default
         if inst_infos is None:
@@ -51,39 +54,43 @@ class GreedyPolicyQActsAny(PathFindQHeur[D, InstanceGrPolQ], ABC):
             self.itr += 1  # TODO make more elegant
             return []
 
-        # get curr edges
-        edges: List[EdgeQ] = [inst.curr_edge for inst in instances]
-
-        # next node
-        nodes_next_l: List[List[Node]] = self.get_next_nodes(instances, [[edge] for edge in edges])
-        nodes_next: List[Node] = [nodes_next_i[0] for nodes_next_i in nodes_next_l]
+        # get curr nodes
+        nodes: List[Node] = [inst.curr_node for inst in instances]
 
         # is solved
         start_time = time.time()
-        self.set_is_solved(nodes_next)
-        for instance, node_next in zip(instances, nodes_next):
-            instance.record_goal(node_next)
+        self.set_is_solved(nodes)
+        for instance, node in zip(instances, nodes):
+            instance.record_goal(node)
+
         self.times.record_time("is_solved", time.time() - start_time)
 
-        start_time = time.time()
-        q_vals_l: List[List[float]] = []
-        for node_next in nodes_next:
-            assert node_next.q_values is not None
-            q_vals_l.append(node_next.q_values[1])
-        self.times.record_time("q_vals", time.time() - start_time)
-
-        # get next idxs
+        # get edge
         start_time = time.time()
         temps: List[float] = [instance.temp for instance in instances]
         eps_l: List[float] = [instance.eps for instance in instances]
+        actions_l: List[List[Action]] = []
+        q_vals_l: List[List[float]] = []
+        for node in nodes:
+            assert node.q_values is not None
+            actions_l.append(node.q_values[0])
+            q_vals_l.append(node.q_values[1])
+
         next_idxs: List[int] = greedy_next_idx(q_vals_l, temps, eps_l)
+
+        edges: List[EdgeQ] = [EdgeQ(node, actions[next_idx], q_vals[next_idx]) for node, actions, q_vals, next_idx in
+                              zip(nodes, actions_l, q_vals_l, next_idxs)]
+
         self.times.record_time("next_idx", time.time() - start_time)
+
+        # next nodes
+        nodes_next_l: List[List[Node]] = self.get_next_nodes(instances, [[edge] for edge in edges])
+        nodes_next: List[Node] = [nodes_next_i[0] for nodes_next_i in nodes_next_l]
 
         # update inst
         start_time = time.time()
-        for instance, node_next, q_vals, next_idx in zip(instances, nodes_next, q_vals_l, next_idxs):
-            assert node_next.q_values is not None
-            instance.curr_edge = EdgeQ(node_next, node_next.q_values[0][next_idx], q_vals[next_idx])
+        for instance, node in zip(instances, nodes_next):
+            instance.curr_node = node
             instance.itr += 1
         self.times.record_time("set_next", time.time() - start_time)