+++ CHANGES.md
@@ -1,8 +1,8 @@
 # Changes
 
 ## 0.2.1
-* Replay buffer added for V
-* HER added for V
+* Replay buffer added
+* HER added
 * add method to Updater to get state actions
 * separate step and step_sync_main method in Updater
 * Explicitly call del and gc during update to better free memory